requirements (1).txt
tensorflow==2.15.0
gradio==3.48.0
numpy==1.24.3
pillow==10.0.0